+++ src/helpers/log_helpers.ts
@@ -11,7 +11,7 @@ const ENABLE_TRACE_LOG = isDebugMode();
 
 // Log prefixes
 export const CLEANUP_FEED = "cleanupFeed()";
-export const GET_FEED = "getFeed()";
+export const TRIGGER_FEED = "triggerFeedUpdate()";
 export const PREP_SCORERS = "prepareScorers()";